+++ src/ResourceManager/Sql/Commands.Sql/Elastic Pools/Services/AzureSqlElasticPoolAdapter.cs
@@ -25,6 +25,7 @@ using System.Linq;
 using Microsoft.Azure.Commands.ResourceManager.Common.Tags;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using DatabaseEdition = Microsoft.Azure.Commands.Sql.Database.Model.DatabaseEdition;
+using System.Globalization;
 
 namespace Microsoft.Azure.Commands.Sql.ElasticPool.Services
 {
@@ -164,12 +165,63 @@ namespace Microsoft.Azure.Commands.Sql.ElasticPool.Services
         /// <returns>A list of Elastic Pool Activities</returns>
         internal IList<AzureSqlElasticPoolActivityModel> GetElasticPoolActivity(string resourceGroupName, string serverName, string poolName)
         {
-            var resp = Communicator.ListActivity(resourceGroupName, serverName, poolName);
+            var activityResp = Communicator.ListActivity(resourceGroupName, serverName, poolName);
+            var operationResp = Communicator.ListOperation(resourceGroupName, serverName, poolName);
 
-            return resp.Select((activity) =>
+            var resp = from activity in activityResp
+                       join operation in operationResp
+                       on activity.OperationId equals new Guid(operation.Name)
+                       select new {
+                           elasticPoolName = activity.ElasticPoolName,
+                           endTime = activity.EndTime,
+                           rgName = resourceGroupName,
+                           errorCode = activity.ErrorCode,
+                           errorMessage = activity.ErrorMessage,
+                           errorSeverity = activity.ErrorSeverity,
+                           operation = activity.Operation,
+                           operationId = activity.OperationId,
+                           percentComplete = activity.PercentComplete,
+                           requestedDatabaseDtuMax = activity.RequestedDatabaseDtuMax,
+                           requestedDatabaseDtuMin = activity.RequestedDatabaseDtuMin,
+                           requestedDtu = activity.RequestedDtu,
+                           requestedElasticPoolName = activity.RequestedElasticPoolName,
+                           requestedStorageLimitInGB = activity.RequestedStorageLimitInGB,
+                           serverName = activity.ServerName,
+                           startTime = activity.StartTime,
+                           state = activity.State,
+                           estimatedCompletionTime = operation.EstimatedCompletionTime,
+                           description = operation.Description,
+                           isCancellable = operation.IsCancellable
+                       };
+
+            IEnumerable<AzureSqlElasticPoolActivityModel> listResponse = resp.Select((r) =>
             {
-                return CreateActivityModelFromResponse(activity);
-            }).ToList();
+                return new AzureSqlElasticPoolActivityModel()
+                {
+                    ElasticPoolName = r.elasticPoolName,
+                    ResourceGroupName = r.rgName,
+                    EndTime = r.endTime,
+                    ErrorCode = r.errorCode,
+                    ErrorMessage = r.errorMessage,
+                    ErrorSeverity = r.errorSeverity,
+                    Operation = r.operation,
+                    OperationId = r.operationId,
+                    PercentComplete = r.percentComplete,
+                    RequestedDatabaseDtuMax = r.requestedDatabaseDtuMax,
+                    RequestedDatabaseDtuMin = r.requestedDatabaseDtuMin,
+                    RequestedDtu = r.requestedDtu,
+                    RequestedElasticPoolName = r.requestedElasticPoolName,
+                    RequestedStorageLimitInGB = r.requestedStorageLimitInGB,
+                    ServerName = r.serverName,
+                    StartTime = r.startTime,
+                    State = r.state,
+                    EstimatedCompletionTime = r.estimatedCompletionTime,
+                    Description = r.description,
+                    IsCancellable = r.isCancellable
+                };
+            });
+
+            return listResponse.ToList();
         }
 
         /// <summary>
@@ -187,6 +239,27 @@ namespace Microsoft.Azure.Commands.Sql.ElasticPool.Services
             {
                 return CreateDatabaseActivityModelFromResponse(activity);
             }).ToList();
+        }
+
+        /// <summary>
+        /// Cancel the elastic pool activity
+        /// </summary>
+        /// <param name="resourceGroupName">The name of the resource group</param>
+        /// <param name="serverName">The name of the Azure Sql Database server</param>
+        /// <param name="elasticPoolName">The name of the elastic pool</param>
+        /// <param name="operationId">The Operation ID</param>
+        /// <returns></returns>
+        internal IEnumerable<AzureSqlElasticPoolActivityModel> CancelElasticPoolActivity(string resourceGroupName, string serverName, string elasticPoolName, Guid? operationId)
+        {
+            if (!operationId.HasValue)
+            {
+                throw new NotSupportedException(string.Format(CultureInfo.InvariantCulture, Microsoft.Azure.Commands.Sql.Properties.Resources.OperationIdRequired));
+            }
+
+            Communicator.CancelOperation(resourceGroupName, serverName, elasticPoolName, operationId.Value);
+
+            // After Cancel event is fired, state will be in 'CancelInProgress' for a while but should expect to finish in a minute
+            return GetElasticPoolActivity(resourceGroupName, serverName, elasticPoolName);
         }
 
         /// <summary>
@@ -243,7 +316,6 @@ namespace Microsoft.Azure.Commands.Sql.ElasticPool.Services
                 StartTime = model.StartTime,
                 State = model.State
             };
-
 
             return activity;
         }