+++ src/ResourceManager/NotificationHubs/Commands.NotificationHubs.Test/ScenarioTests/NHServiceTests.cs
@@ -39,15 +39,17 @@ namespace Commands.NotificationHubs.Test
             RunPowerShellTest("Test-CRUDNamespaceAuth");
         }
 
-        [Fact]
+        [Fact(Skip = "Need service team to re-record test after changes to the ClientRuntime.")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait("Re-record", "ClientRuntime changes")]
         public void TestCRUDNotificationHub()
         {
             RunPowerShellTest("Test-CRUDNotificationHub");
         }
 
-        [Fact]
+        [Fact(Skip = "Need service team to re-record test after changes to the ClientRuntime.")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait("Re-record", "ClientRuntime changes")]
         public void TestCRUDNHAuth()
         {
             RunPowerShellTest("Test-CRUDNHAuth");