+++ src/ResourceManager/Sql/ChangeLog.md
@@ -20,6 +20,7 @@
 ## Current Release
 * Add new cmdlet 'Stop-AzureRmSqlElasticPoolActivity' to support canceling the asynchronous operations on elastic pool
 * Update the response for cmdlets Get-AzureRmSqlDatabaseActivity and Get-AzureRmSqlElasticPoolActivity to reflect more information in the response
+* Updated to the latest version of the Azure ClientRuntime
 
 ## Version 4.3.1
 * Fix issue with Default Resource Group in CloudShell